CRUD data from SQLITE

+++ www/views/restaurants/restaurants.js
@@ -1,6 +1,6 @@
 angular.module('App')
 .controller('RestaurantsController', function ($scope, $http,
-  Restaurant,$ionicModal) {
+  Restaurant,$ionicModal,$ionicPopup) {
 
   $scope.page = 0;
   $scope.total = 1;
@@ -29,9 +29,23 @@ angular.module('App')
   $scope.getRestaurants();
 
   $scope.hapus = function(id){
-    Restaurant.deleteResto(id);
-    console.log('menghapus data id:',id);
-    $scope.getRestaurants();
+    $ionicPopup.confirm({
+     title: 'Delete Resto',
+     template: 'Are you sure you want to delete resto?'
+    }).then(function(res) {
+     if(res) {
+        Restaurant.deleteResto(id);
+        console.log('menghapus data id:',id);
+        $scope.getRestaurants();
+     } else {
+       console.log('You are not sure');
+     }
+   });
+
+
+
+
+   
   }
 
   $scope.modal;
@@ -49,13 +63,35 @@ angular.module('App')
       );
 
   $scope.add = function(){
+    $scope.save = true;
     $scope.modal.show();
   }
 
   $scope.simpan = function(restaurant){
-    Restaurant.addResto(restaurant);
+    if($scope.save){
+      Restaurant.addResto(restaurant);
+    }else{
+      Restaurant.updateResto(restaurant);
+    }
+    restaurant.name='';
+    restaurant.address='';
+    restaurant.city='';
     $scope.modal.hide();
     $scope.getRestaurants();
+  }
+
+  $scope.save = true;
+
+  $scope.restaurant = {id:'',name:'',address:'',city:'',image_url:''}
+  $scope.update = function(restaurant){
+    $scope.restaurant.name = restaurant.name;
+    $scope.restaurant.address = restaurant.address;
+    $scope.restaurant.city = restaurant.city;
+    $scope.restaurant.image_url = restaurant.image_url;
+    $scope.restaurant.id = restaurant.id;
+
+    $scope.save = false;
+    $scope.modal.show();
   }
 
 });
